+++ src/event.ts
@@ -10,3 +10,5 @@ export const videoTimeUpdate = createEvent("Video time update");
 export const toggleShowFullSubTranslatePopup = createEvent("Toggle show full sub translate popup");
 
 export const toggleAutoPause = createEvent("Toggle auto pause");
+
+export const setSubsFontSize = createEvent("Set subs font size");